fix: 추가 정보 입력 palette, margin 완료
