+++ app/structure/lib/components/data_title.dart
@@ -1,5 +1,4 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_screenutil/flutter_screenutil.dart';
 
 class DataTitle extends StatelessWidget {
   const DataTitle({super.key, required this.korText, required this.engText});
@@ -9,23 +8,20 @@ class DataTitle extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return Container(
-      margin: EdgeInsets.symmetric(horizontal: 40.w),
-      alignment: Alignment.centerLeft,
-      child: Row(
-        children: [
-          Text(
-            '$engText ',
-            textAlign: TextAlign.left,
-            style: const TextStyle(fontWeight: FontWeight.bold),
-          ),
-          Text(
-            korText,
-            textAlign: TextAlign.left,
-            style: const TextStyle(color: Colors.grey),
-          ),
-        ],
-      ),
+    return Row(
+      mainAxisAlignment: MainAxisAlignment.start,
+      children: [
+        Text(
+          '$engText ',
+          textAlign: TextAlign.left,
+          style: const TextStyle(fontWeight: FontWeight.bold),
+        ),
+        Text(
+          korText,
+          textAlign: TextAlign.left,
+          style: const TextStyle(color: Colors.grey),
+        ),
+      ],
     );
   }
 }
