+++ app/structure/lib/screen/data_management/researcher/insertion_lab_data_screen.dart
@@ -32,7 +32,6 @@ class _InsertionLabDataScreenState extends State<InsertionLabDataScreen> {
         FocusScope.of(context).unfocus();
       },
       child: Scaffold(
-        backgroundColor: Colors.white,
         appBar: CustomAppBar(
           title: insertionLabDataViewModel.title,
           backButton: true,
@@ -41,127 +40,136 @@ class _InsertionLabDataScreenState extends State<InsertionLabDataScreen> {
         body: Stack(
           children: [
             SingleChildScrollView(
-              child: Column(
-                mainAxisSize: MainAxisSize.min,
-                children: [
-                  DataField(
-                    mainText: 'L ',
-                    subText: '명도',
-                    controller: insertionLabDataViewModel.l,
-                    onChangeFunc: (_) => insertionLabDataViewModel.inputCheck(),
-                  ),
-                  SizedBox(height: 32.h),
+              child: Container(
+                margin: EdgeInsets.symmetric(horizontal: 40.w),
+                child: Column(
+                  mainAxisSize: MainAxisSize.min,
+                  children: [
+                    SizedBox(height: 24.h),
 
-                  DataField(
-                    mainText: 'a ',
-                    subText: '적색도',
-                    unit: '',
-                    controller: insertionLabDataViewModel.a,
-                    onChangeFunc: (_) => insertionLabDataViewModel.inputCheck(),
-                  ),
-                  SizedBox(height: 32.h),
+                    DataField(
+                      mainText: 'L ',
+                      subText: '명도',
+                      controller: insertionLabDataViewModel.l,
+                      onChangeFunc: (_) =>
+                          insertionLabDataViewModel.inputCheck(),
+                    ),
+                    SizedBox(height: 32.h),
 
-                  DataField(
-                    mainText: 'b ',
-                    subText: '황색도',
-                    unit: '',
-                    controller: insertionLabDataViewModel.b,
-                    onChangeFunc: (_) => insertionLabDataViewModel.inputCheck(),
-                  ),
-                  SizedBox(height: 32.h),
+                    DataField(
+                      mainText: 'a ',
+                      subText: '적색도',
+                      unit: '',
+                      controller: insertionLabDataViewModel.a,
+                      onChangeFunc: (_) =>
+                          insertionLabDataViewModel.inputCheck(),
+                    ),
+                    SizedBox(height: 32.h),
 
-                  DataField(
-                    mainText: 'DL ',
-                    subText: '육즙감량',
-                    unit: '%',
-                    controller: insertionLabDataViewModel.dl,
-                    isPercent: true,
-                    onChangeFunc: (_) => insertionLabDataViewModel.inputCheck(),
-                  ),
-                  SizedBox(height: 32.h),
+                    DataField(
+                      mainText: 'b ',
+                      subText: '황색도',
+                      unit: '',
+                      controller: insertionLabDataViewModel.b,
+                      onChangeFunc: (_) =>
+                          insertionLabDataViewModel.inputCheck(),
+                    ),
+                    SizedBox(height: 32.h),
 
-                  DataField(
-                    mainText: 'CL ',
-                    subText: '가열감량',
-                    unit: '%',
-                    controller: insertionLabDataViewModel.cl,
-                    isPercent: true,
-                    onChangeFunc: (_) => insertionLabDataViewModel.inputCheck(),
-                  ),
-                  SizedBox(height: 32.h),
+                    DataField(
+                      mainText: 'DL ',
+                      subText: '육즙감량',
+                      unit: '%',
+                      controller: insertionLabDataViewModel.dl,
+                      isPercent: true,
+                      onChangeFunc: (_) =>
+                          insertionLabDataViewModel.inputCheck(),
+                    ),
+                    SizedBox(height: 32.h),
 
-                  DataField(
-                    mainText: 'RW ',
-                    subText: '압착감량',
-                    unit: '%',
-                    controller: insertionLabDataViewModel.rw,
-                    isPercent: true,
-                    onChangeFunc: (_) => insertionLabDataViewModel.inputCheck(),
-                  ),
-                  SizedBox(height: 32.h),
+                    DataField(
+                      mainText: 'CL ',
+                      subText: '가열감량',
+                      unit: '%',
+                      controller: insertionLabDataViewModel.cl,
+                      isPercent: true,
+                      onChangeFunc: (_) =>
+                          insertionLabDataViewModel.inputCheck(),
+                    ),
+                    SizedBox(height: 32.h),
 
-                  DataField(
-                    mainText: 'pH ',
-                    subText: '산도',
-                    controller: insertionLabDataViewModel.ph,
-                    onChangeFunc: (_) => insertionLabDataViewModel.inputCheck(),
-                  ),
-                  SizedBox(height: 32.h),
+                    DataField(
+                      mainText: 'RW ',
+                      subText: '압착감량',
+                      unit: '%',
+                      controller: insertionLabDataViewModel.rw,
+                      isPercent: true,
+                      onChangeFunc: (_) =>
+                          insertionLabDataViewModel.inputCheck(),
+                    ),
+                    SizedBox(height: 32.h),
 
-                  DataField(
-                    mainText: 'WBSF ',
-                    subText: '전단가',
-                    unit: 'kgf',
-                    controller: insertionLabDataViewModel.wbsf,
-                    onChangeFunc: (_) => insertionLabDataViewModel.inputCheck(),
-                  ),
-                  SizedBox(height: 32.h),
+                    DataField(
+                      mainText: 'pH ',
+                      subText: '산도',
+                      controller: insertionLabDataViewModel.ph,
+                      onChangeFunc: (_) =>
+                          insertionLabDataViewModel.inputCheck(),
+                    ),
+                    SizedBox(height: 32.h),
 
-                  DataField(
-                    mainText: '카텝신활성도',
-                    subText: '',
-                    controller: insertionLabDataViewModel.ct,
-                    onChangeFunc: (_) => insertionLabDataViewModel.inputCheck(),
-                  ),
-                  SizedBox(height: 32.h),
+                    DataField(
+                      mainText: 'WBSF ',
+                      subText: '전단가',
+                      unit: 'kgf',
+                      controller: insertionLabDataViewModel.wbsf,
+                      onChangeFunc: (_) =>
+                          insertionLabDataViewModel.inputCheck(),
+                    ),
+                    SizedBox(height: 32.h),
 
-                  DataField(
-                    mainText: 'MFI ',
-                    subText: '근소편화지수',
-                    controller: insertionLabDataViewModel.mfi,
-                    onChangeFunc: (_) => insertionLabDataViewModel.inputCheck(),
-                  ),
-                  DataField(
-                      isFinal: 1,
+                    DataField(
+                      mainText: '카텝신활성도',
+                      subText: '',
+                      controller: insertionLabDataViewModel.ct,
+                      onChangeFunc: (_) =>
+                          insertionLabDataViewModel.inputCheck(),
+                    ),
+                    SizedBox(height: 32.h),
+
+                    DataField(
+                      mainText: 'MFI ',
+                      subText: '근소편화지수',
+                      controller: insertionLabDataViewModel.mfi,
+                      onChangeFunc: (_) =>
+                          insertionLabDataViewModel.inputCheck(),
+                    ),
+                    SizedBox(height: 32.h),
+
+                    DataField(
                       mainText: 'Collagen ',
                       subText: '콜라겐',
-                      controller:
-                          context.read<InsertionLabDataViewModel>().collagen),
-                  SizedBox(height: 32.h),
-
-                  DataField(
-                    mainText: 'Collagen ',
-                    subText: '콜라겐',
-                    controller: insertionLabDataViewModel.collagen,
-                    onChangeFunc: (_) => insertionLabDataViewModel.inputCheck(),
-                  ),
-                  SizedBox(height: 64.h),
-
-                  // 저장 버튼
-                  Container(
-                    margin: EdgeInsets.fromLTRB(40.w, 0, 40.w, 40.w),
-                    child: MainButton(
-                      onPressed: insertionLabDataViewModel.inputComplete
-                          ? () async =>
-                              insertionLabDataViewModel.saveData(context)
-                          : null,
-                      text: '저장',
-                      width: double.infinity,
-                      height: 96.h,
-                      mode: 1,
+                      controller: insertionLabDataViewModel.collagen,
+                      onChangeFunc: (_) =>
+                          insertionLabDataViewModel.inputCheck(),
                     ),
-                  ),
-                ],
+                    SizedBox(height: 64.h),
+
+                    // 저장 버튼
+                    Container(
+                      margin: EdgeInsets.only(bottom: 40.h),
+                      child: MainButton(
+                        width: double.infinity,
+                        height: 96.h,
+                        text: '저장',
+                        onPressed: insertionLabDataViewModel.inputComplete
+                            ? () async =>
+                                insertionLabDataViewModel.saveData(context)
+                            : null,
+                      ),
+                    ),
+                  ],
+                ),
               ),
             ),
             insertionLabDataViewModel.isLoading
